+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [{VERSION}] for KSP {KSPVERSION} - {DATE}
+### Added
+- Added (english) KSPedia. Fixes #17
+
+### Fixed
+- fixed the DPAI window remaining visible on top of KSPedia. Fixes #27
 
 ## [6.12.0.0] for KSP 1.12.5 - 2025-05-15
 ### Added